requirements.txt
pytest==8.1.2
selenium==4.20.0
pytest-html==4.1.1
pytest-xdist==3.5.0
PyYAML==6.0.1